Restructured the plugin for performance and maintainability.
